+++ lib/Makefile
@@ -10,6 +10,7 @@ LIBCOBJ += vgs_math.o
 LIBCOBJ += vgs_ctype.o
 LIBCOBJ += vgs_save.o
 LIBCOBJ += vgs_key.o
+LIBCOBJ += vgs_bmpfont.o
 
 all: libc.a liblog.a
 	cp -p vgs_io.h ../src
